requirements.txt
crewai
crewai-tools
python-dotenv
yfinance
openai
newsapi-python
nltk
transformers
torch
fredapi